Remove redundant cast

+++ warp/src/test/java/me/sparky983/warp/internal/deserializers/ByteDeserializerTest.java
@@ -43,6 +43,6 @@ class ByteDeserializerTest {
 
     final byte result = BYTE.deserialize(node);
 
-    assertEquals((byte) value, result);
+    assertEquals(value, result);
   }
 }
